Back/check_list/src/requirements.txt
asgiref==3.4.1
Command==0.1.0
Django
django-cors-headers==3.10.0
django-dbbackup==4.0.0b0
django-guardian==2.4.0
django-ipware==4.0.2
django-phonenumber-field==6.0.0
django-service-objects==0.7.0
django-structlog==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
numpy
openpyxl==3.0.9
pandas==1.4.1
phonenumbers==8.12.39
Pillow==9.0.1
psycopg2==2.9.1
PyJWT==1.7.1
python-dateutil
pytz==2021.1
pywin32==303
six==1.16.0
sqlparse==0.4.2
structlog==21.5.0
xlwings